requirements.txt
# SwiftVisa Backend - Python Dependencies
# Generated: November 29, 2025
# Python Version: 3.12+

# Core Web Framework
fastapi==0.121.0
uvicorn[standard]==0.34.0
pydantic==2.10.4
python-multipart==0.0.20

# Environment & Configuration
python-dotenv==1.0.1

# LangChain & AI/ML
langchain==0.3.13
langchain-community==0.3.13
langchain-chroma==0.2.0
langchain-huggingface==0.1.2
langchain-openai==0.2.14
langchain-google-genai==2.0.8

# Vector Store & Embeddings
chromadb==0.5.23
sentence-transformers==3.3.1
huggingface-hub==0.27.0

# OpenAI & Google AI
openai==1.59.3
google-generativeai==0.8.3

# Data Processing
numpy==2.2.0
pandas==2.2.3

# NLP
nltk==3.9.1

# HTTP Client
requests==2.32.3
httpx==0.28.1

# Utilities
tqdm==4.67.1
python-dateutil==2.9.0.post0

# CORS Middleware
starlette==0.45.2

# Optional - For production deployment
gunicorn==23.0.0

# Optional - For monitoring and logging
python-json-logger==3.2.1
